add: implemented faculty and guardian signup activity

+++ app/src/main/java/com/stec/srms/database/StudentDBHandler.java
@@ -288,13 +288,14 @@ public class StudentDBHandler extends Database {
     }
 
     public boolean deletePendingStudent(int userId) {
+        int accountId = getAccountType("pendingStudent").accountId;
         SQLiteDatabase db = null;
         try {
             db = this.getWritableDatabase();
             db.beginTransaction();
 
             db.delete("pending_students", "userId = " + userId, null);
-            db.delete("pending_verifications", "userId = " + userId, null);
+            db.delete("pending_verifications", "accountId = " + accountId + " AND userId = " + userId, null);
 
             db.setTransactionSuccessful();
             return true;
